Implement SSR for ProgressBar
