+++ testing/tests/DevExpress.serverSide/progressBar.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets/progressBar.markup.tests.js");
