+++ testing/tests/DevExpress.ui.widgets/progressBar.tests.js
@@ -25,8 +25,6 @@ function toSelector(text) {
 
 var PROGRESSBAR_CLASS = "dx-progressbar",
     PROGRESSBAR_CONTAINER_CLASS = "dx-progressbar-container",
-    PROGRESSBAR_RANGE_CLASS = "dx-progressbar-range",
-    PROGRESSBAR_WRAPPER_CLASS = "dx-progressbar-wrapper",
     PROGRESSBAR_STATUS_CLASS = "dx-progressbar-status",
     PROGRESSBAR_INDETERMINATE_SEGMENT_CONTAINER = "dx-progressbar-animating-container",
     PROGRESSBAR_INDETERMINATE_SEGMENT = "dx-progressbar-animating-segment";
@@ -35,18 +33,6 @@ QUnit.module("default", {
     beforeEach: function() {
         this.$element = $("#progressbar");
     }
-});
-
-QUnit.test("rendered markup", function(assert) {
-    assert.expect(5);
-
-    var $progressBar = this.$element.dxProgressBar();
-
-    assert.ok($progressBar.hasClass(PROGRESSBAR_CLASS), "ProgressBar initialized");
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_CONTAINER_CLASS)).length, 1, "Container has been created");
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_RANGE_CLASS)).length, 1, "Range has been created");
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_WRAPPER_CLASS)).length, 1, "Wrapper div has been created");
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_STATUS_CLASS)).length, 1, "Status div has been created");
 });
 
 QUnit.test("onContentReady fired after the widget is fully ready", function(assert) {
@@ -69,30 +55,12 @@ QUnit.module("options", {
     }
 });
 
-QUnit.test("showStatus test", function(assert) {
-    assert.expect(2);
-
-    var $progressBar = this.$element.dxProgressBar({
-            showStatus: false
-        }),
-        progressBar = $progressBar.dxProgressBar("instance");
-
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_STATUS_CLASS)).length, 0, "Status div hasn't been created");
-
-    progressBar.option("showStatus", true);
-    assert.equal($progressBar.find(toSelector(PROGRESSBAR_STATUS_CLASS)).length, 1, "Status div has been added");
-});
-
-QUnit.test("value display in status", function(assert) {
-    assert.expect(2);
-
+QUnit.test("correct value display in status after option changed", function(assert) {
     var $progressBar = this.$element.dxProgressBar({
             value: 10
         }),
         progressBar = $progressBar.dxProgressBar("instance"),
         $status = $progressBar.find(toSelector(PROGRESSBAR_STATUS_CLASS));
-
-    assert.equal($status.text(), "Progress: " + progressBar.option("value") + "%", "status text is right");
 
     progressBar.option("value", 30);
     assert.equal($status.text(), "Progress: " + progressBar.option("value") + "%", "status text has been change right");
@@ -180,30 +148,6 @@ QUnit.test("complete option changed", function(assert) {
     assert.equal(secondCompleteActionFired, 1, "second CompleteActionFired is fired");
 });
 
-QUnit.test("appropriate class should be added depending on the 'statusPosition' option", function(assert) {
-    var possiblePositions = [
-        "left", "right", "bottom left", "bottom right", "bottom center", "top left", "top right", "top center"
-    ];
-
-    var $progressBar = this.$element.dxProgressBar({}),
-        progressBar = $progressBar.dxProgressBar("instance"),
-        $wrapper = $progressBar.find("." + PROGRESSBAR_WRAPPER_CLASS);
-
-    $.each(possiblePositions, function(_, position) {
-        progressBar.option("statusPosition", position);
-
-        var splitPosition = position.split(" "),
-            className = "dx-position-" + splitPosition[0];
-
-        if(splitPosition[1]) {
-            className += "-" + splitPosition[1];
-        }
-
-        assert.ok($wrapper.hasClass(className), "wrapper correct class for the '" + position + "' position");
-    });
-});
-
-
 QUnit.module("states", {
     beforeEach: function() {
         this.$element = $("#progressbar");
@@ -239,37 +183,15 @@ QUnit.test("render indeterminate state", function(assert) {
     assert.equal($progressBar.find(toSelector(PROGRESSBAR_CONTAINER_CLASS)).is(":visible"), false, "progressbar container does not attached");
 });
 
-QUnit.test("render indeterminate state with default option segments count", function(assert) {
-    var $progressBar = this.$element.dxProgressBar({
-            value: undefined,
-            showStatus: false
-        }),
-        progressBar = $progressBar.dxProgressBar("instance");
-
-    var renderedIndeterminateSegmentsCount = $progressBar.find(toSelector(PROGRESSBAR_INDETERMINATE_SEGMENT)).length,
-        defaultSegmentCount = progressBar.option("_animatingSegmentCount");
-
-    assert.equal(renderedIndeterminateSegmentsCount, defaultSegmentCount, "dxProgressBar have been created with correct segment count");
-});
-
 QUnit.module("aria accessibility");
 
-QUnit.test("aria role", function(assert) {
-    var $element = $("#progressbar").dxProgressBar({});
-    assert.equal($element.attr("role"), "progressbar", "aria role is correct");
-});
-
-QUnit.test("aria properties", function(assert) {
+QUnit.test("aria properties after options changed", function(assert) {
     var $element = $("#progressbar").dxProgressBar({
             min: 32,
             max: 137,
             value: 58
         }),
         instance = $element.dxProgressBar("instance");
-
-    assert.equal($element.attr("aria-valuemin"), 32, "min value is correct");
-    assert.equal($element.attr("aria-valuemax"), 137, "max value is correct");
-    assert.equal($element.attr("aria-valuenow"), 58, "current value is correct");
 
     instance.option({
         min: 33,
